Move from jest to vitest
Change the test framework to use vitest instead of jest.
This is faster and provides a simpler configuration when using
typescript
