--- jest.config.json
@@ -1,12 +1,0 @@
-{
-    "moduleFileExtensions": [
-        "js"
-    ],
-    "testEnvironment": "node",
-    "transform": {},
-    "setupFiles": [
-        "./lib/test/jest.setup-file.js"
-    ],
-    "collectCoverage": true,
-    "verbose": true
-}
