+++ package.json
@@ -29,7 +29,7 @@
     },
     "scripts": {
         "build": "tsc -b",
-        "test": "tsc -b test && node --experimental-vm-modules node_modules/jest/bin/jest.js --ci",
+        "test": "tsc -b test && vitest --run",
         "lint": "eslint .",
         "format": "prettier './(src|test|bench|examples)/**/*.ts' --write",
         "bench": "tsc -b bench && node ./lib/bench/index.js",
@@ -47,17 +47,17 @@
     "devDependencies": {
         "@eslint/js": "9.9.1",
         "@types/benchmark": "2.1.5",
-        "@types/jest": "29.5.12",
+        "@vitest/coverage-v8": "3.0.7",
         "benchmark": "2.1.4",
-        "eslint": "9.9.1",
+        "eslint": "8.57.1",
         "eslint-config-prettier": "9.1.0",
-        "eslint-plugin-jest": "28.8.1",
         "eslint-plugin-prettier": "5.2.1",
         "eslint-plugin-security": "3.0.1",
-        "jest": "29.7.0",
+        "eslint-plugin-vitest": "0.5.4",
         "prettier": "3.3.3",
         "rimraf": "6.0.1",
         "typescript": "5.5.4",
-        "typescript-eslint": "8.3.0"
+        "typescript-eslint": "8.26.0",
+        "vitest": "3.0.7"
     }
 }
