+++ test/util.ts
@@ -1,6 +1,8 @@
 // Copyright (c) 2023 Cloudflare, Inc.
 // Licensed under the Apache-2.0 license found in the LICENSE file or at https://opensource.org/licenses/Apache-2.0
 
+import { expect } from 'vitest';
+
 import type { TokenTypeEntry } from '../src/index.js';
 
 export function hexToString(x: string): string {
